moved methods to utils, finished the refactoring process
